requirements.txt
fastapi==0.133.1
uvicorn==0.38.0
mysql-connector-python==9.5.0
requests==2.32.5
